Make Bosen work with DNS names in machinefile
Bosen binds to the IP addresses listed in machinefile. Changing IP
addresses to DNS names breaks Bosen since ZMQ binds only to IP addresses.
There is no requirement for Bosen to bind to the specific IP addresses
listed in machinefile, only the port number is important.  This patch
changes Bosen to bind to tcp://*.port-number-from-machine-file.  With this
change Bosen works with DNS names in the machinefile.  It is
backward-compatible -- listing IP addresses in machinefile works as well.